eliminate redundancies in Ldap classes
