+++ roles/middleware/files/FWO.Middleware.Server/Ldap.cs
@@ -5,61 +5,12 @@ using System.Linq;
 using System.Collections.Generic;
 using System.Net.Security;
 using System.Security.Cryptography.X509Certificates;
-using System.Text.Json.Serialization;
 using FWO.Api.Data;
 
 namespace FWO.Middleware.Server
 {
-    public class Ldap
+    public class Ldap : LdapConnectionBase
     {
-        // The following properties are retrieved from the database api:
-        // ldap_server ldap_port ldap_search_user ldap_tls ldap_tenant_level ldap_connection_id ldap_search_user_pwd ldap_searchpath_for_users ldap_searchpath_for_roles    
-
-        [JsonPropertyName("ldap_connection_id")]
-        public int Id { get; set; }
-
-        [JsonPropertyName("ldap_server")]
-        public string Address { get; set; }
-
-        [JsonPropertyName("ldap_port")]
-        public int Port { get; set; }
-
-        [JsonPropertyName("ldap_type")]
-        public int Type { get; set; }
-
-        [JsonPropertyName("ldap_pattern_length")]
-        public int PatternLength { get; set; }
-
-        [JsonPropertyName("ldap_search_user")]
-        public string SearchUser { get; set; }
-
-        [JsonPropertyName("ldap_tls")]
-        public bool Tls { get; set; }
-
-        [JsonPropertyName("ldap_tenant_level")]
-        public int TenantLevel { get; set; }
-
-        [JsonPropertyName("ldap_search_user_pwd")]
-        public string SearchUserPwd { get; set; }
-
-        [JsonPropertyName("ldap_searchpath_for_users")]
-        public string UserSearchPath { get; set; }
-
-        [JsonPropertyName("ldap_searchpath_for_roles")]
-        public string RoleSearchPath { get; set; }
-
-        [JsonPropertyName("ldap_searchpath_for_groups")]
-        public string GroupSearchPath { get; set; }
-
-        [JsonPropertyName("ldap_write_user")]
-        public string WriteUser { get; set; }
-
-        [JsonPropertyName("ldap_write_user_pwd")]
-        public string WriteUserPwd { get; set; }
-
-        [JsonPropertyName("tenant_id")]
-        public int? TenantId { get; set; }
-
         private const int timeOutInMs = 3000;
 
         /// <summary>
@@ -81,31 +32,6 @@ namespace FWO.Middleware.Server
             {
                 throw new Exception($"Error while trying to reach LDAP server {Address}:{Port}", exception);
             }
-        }
-
-        public string Host()
-        {
-            return Address + ":" + Port;
-        }
-
-        public bool IsWritable()
-        {
-            return (WriteUser != null && WriteUser != "");
-        }
-
-        public bool HasGroupHandling()
-        {
-            return (GroupSearchPath != null && GroupSearchPath != "");
-        }
-
-        public bool HasRoleHandling()
-        {
-            return (RoleSearchPath != null && RoleSearchPath != "");
-        }
-
-        public bool IsInternal()
-        {
-            return ((new DistName(UserSearchPath)).IsInternal());
         }
 
         private string getUserSearchFilter(string searchPattern)
